refactor: Create the cdc components entry point.
