+++ integration-libs/cdc/components/cdc-components.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { GigyaRaasModule } from '../components/gigya-raas/gigya-raas.module';
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    GigyaRaasModule
+  ]
+})
+export class CdcComponentsModule { }
