+++ integration-libs/cdc/components/gigya-raas/gigya-raas.component.scss
