+++ integration-libs/cdc/components/gigya-raas/gigya-raas.component.ts
@@ -9,9 +9,9 @@ import { BaseSiteService, LanguageService, WindowRef } from '@spartacus/core';
 import { CmsComponentData } from '@spartacus/storefront';
 import { Observable } from 'rxjs';
 import { distinctUntilChanged, take, tap } from 'rxjs/operators';
-import { CdcJsService } from '../../auth/facade/cdc-js.service';
-import { CdcConfig } from '../../../root/config';
-import { GigyaRaasComponentData } from '../cms.model';
+import { CdcJsService } from '../../src/auth/facade/cdc-js.service';
+import { CdcConfig } from '../../root/config';
+import { GigyaRaasComponentData } from '../../src/cms-components/cms.model';
 
 @Component({
   selector: 'cx-gigya-raas',
