+++ integration-libs/cdc/components/index.ts
@@ -1,0 +1,2 @@
+export * from './cdc-components.module';
+export * from './gigya-raas/index';
