+++ projects/storefrontapp/tsconfig.app.prod.json
@@ -97,6 +97,7 @@
       "@spartacus/tracking/tms/gtm": ["dist/tracking/tms/gtm"],
       "@spartacus/tracking/tms": ["dist/tracking/tms"],
       "@spartacus/cdc": ["dist/cdc"],
+      "@spartacus/cdc/components": ["dist/cdc/components"],
       "@spartacus/cds": ["dist/cds"],
       "@spartacus/assets": ["dist/assets"],
       "@spartacus/core": ["dist/core"],
