+++ projects/storefrontapp/tsconfig.server.json
@@ -135,6 +135,10 @@
       ],
       "@spartacus/tracking/tms": ["../../feature-libs/tracking/tms/public_api"],
       "@spartacus/cdc": ["../../integration-libs/cdc/public_api"],
+      "@spartacus/cdc/components": [
+        "../../integration-libs/cdc/components/public_api"
+      ],
+
       "@spartacus/cds": ["../../integration-libs/cds/public_api"],
       "@spartacus/assets": ["../../projects/assets/src/public_api"],
       "@spartacus/core": ["../../projects/core/public_api"],
